Updating terminal messages for recheck subcommand

+++ torrentfile/commands.py
@@ -166,13 +166,13 @@ def recheck(args: list) -> list:
     logger.debug("Validating %s <---------------> %s contents", metafile, content)
     msg = f"Rechecking  {metafile} ..."
     halfterm = shutil.get_terminal_size().columns / 2
-    padding = (halfterm - (len(msg) / 2)) * " "
+    padding = int(halfterm - (len(msg) / 2)) * " "
     print(padding + msg)
     checker = Checker(metafile, content)
     logger.debug("Completed initialization of the Checker class")
     result = checker.results()
     message = f"{content} <- {result}% -> {metafile}"
-    padding = (halfterm - (len(message) / 2)) * " "
+    padding = int(halfterm - (len(message) / 2)) * " "
     sys.stdout.write(padding + message + "\n")
     sys.stdout.flush()
     return result
